+++ src/tests/orga/views/test_orga_views_cfp.py
@@ -361,7 +361,7 @@ def test_can_add_simple_question(orga_client, event):
             "variant": "string",
             "active": True,
             "help_text_0": "Answer if you want to reach the other side!",
-            "question_required": QuestionRequired.NONE,
+            "question_required": QuestionRequired.OPTIONAL,
         },
         follow=True,
     )
@@ -400,7 +400,7 @@ def test_can_add_choice_question(orga_client, event):
             "form-1-answer_0": "European",
             "form-2-id": "",
             "form-2-answer_0": "",
-            "question_required": QuestionRequired.NONE,
+            "question_required": QuestionRequired.OPTIONAL,
         },
         follow=True,
     )
@@ -441,7 +441,7 @@ def test_can_edit_choice_question(orga_client, event, choice_question):
             "form-2-DELETE": "on",
             "form-3-id": "",
             "form-3-answer_0": "",
-            "question_required": QuestionRequired.NONE,
+            "question_required": QuestionRequired.OPTIONAL,
         },
         follow=True,
     )
@@ -551,9 +551,9 @@ def test_can_remind_answered_submission_question(
     count,
 ):
     with scope(event=event):
-        from pretalx.submission.models.question import Answer, QuestionRequired
+        from pretalx.submission.models.question import Answer
 
-        question.question_required = QuestionRequired.REQUIRE
+        question.question_required = QuestionRequired.REQUIRED
         question.deadline = None
         question.save()
         original_count = QueuedMail.objects.count()